Crate tree_sitter_wasm_build_tool

tree-sitter-wasm-build-tool

A crate to easily allow tree-sitter parsers to compile to Rust’s wasm32-unknown-unknown target.

Currently, this only works with parsers that do not make use of an external C++ scanner.

It is meant to be used in build scripts, typically only if some wasm or c2rust feature is enabled.

For example:

let src_dir = std::path::Path::new("src");
let mut c_config = cc::Build::new();
c_config.include(src_dir);
c_config
    .flag("-Wno-unused-parameter")
    .flag("-Wno-unused-but-set-variable")
    .flag("-Wno-trigraphs");
let parser_path = src_dir.join("parser.c");
c_config.file(&parser_path);

#[cfg(feature = "c2rust")]
tree_sitter_wasm_build_tool::add_wasm_headers(&mut c_config).unwrap();

c_config.compile("parser");
println!("cargo:rerun-if-changed={}", parser_path.to_str().unwrap());

The only public function is add_wasm_headers. See its documentation for some more information.

Functions

add_wasm_headers

Add the required C header files to the given instance of cc::Build, to allow compilation of tree-sitter parsers to the wasm32-unknown-unknown target.